src/main/python/airflow/fetch_guides/gcp_fetch/requirements.txt
google-cloud-storage==1.42.0
requests==2.26.0
beautifulsoup4==4.10.0
google-cloud-secret-manager==2.7.0
